ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
langchain==0.0.339
openai==1.3.5
python-dotenv==1.0.0
PyPDF2==3.0.1
